# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "O2geeks Headless CMS",
    "description": "\n**Backend API for the O2geeks Branding System**\n\nThis API powers the decoupled content management backend. It provides robust RESTful endpoints for delivering and managing multi-format content.\n\n### Key Features:\n- **Authentication**: JWT-based secure access with Refresh Tokens.\n- **RBAC**: Strict Role-Based Access Control (Super Admin, Admin, Editor, User).\n- **Rate Limiting**: Distributed Redis-backed abuse protection.\n- **Media**: Dual-provider (Local + AWS S3) Multipart/Form-Data uploads.\n- **Engagement**: Zero-N+1 likes, comments, and favorites.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/news": {
      "get": {
        "tags": [
          "News"
        ],
        "summary": "List News",
        "description": "List all news articles with pagination.\n\n- Public endpoint\n- Filterable by status and is_featured flag",
        "operationId": "list_news_api_v1_news_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 10,
              "title": "Per Page"
            },
            "description": "Items per page"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search term for headline or summary",
              "title": "Search"
            },
            "description": "Search term for headline or summary"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ContentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by publication status",
              "title": "Status"
            },
            "description": "Filter by publication status"
          },
          {
            "name": "is_featured",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter featured news only",
              "title": "Is Featured"
            },
            "description": "Filter featured news only"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sort field",
              "title": "Sort By"
            },
            "description": "Sort field"
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sort order: asc or desc",
              "title": "Sort Order"
            },
            "description": "Sort order: asc or desc"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_NewsOut_"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g., Validation Error or Duplicate Data)"
          },
          "401": {
            "description": "Unauthorized (Missing or Invalid JWT Token)"
          },
          "403": {
            "description": "Forbidden (Insufficient Permissions or Deactivated Account)"
          },
          "404": {
            "description": "Not Found (The requested resource does not exist)"
          },
          "429": {
            "description": "Too Many Requests (Rate limit exceeded for this endpoint)"
          },
          "500": {
            "description": "Internal Server Error (Unexpected system failure)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "News"
        ],
        "summary": "Create News",
        "description": "Create a new news article with optional cover image upload.\n\n- Admin-only endpoint\n- Slug auto-generated from headline",
        "operationId": "create_news_api_v1_news_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_news_api_v1_news_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g., Validation Error or Duplicate Data)"
          },
          "401": {
            "description": "Unauthorized (Missing or Invalid JWT Token)"
          },
          "403": {
            "description": "Forbidden (Insufficient Permissions or Deactivated Account)"
          },
          "404": {
            "description": "Not Found (The requested resource does not exist)"
          },
          "429": {
            "description": "Too Many Requests (Rate limit exceeded for this endpoint)"
          },
          "500": {
            "description": "Internal Server Error (Unexpected system failure)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/news/{slug}": {
      "get": {
        "tags": [
          "News"
        ],
        "summary": "Get News",
        "description": "Get a single news article by its URL slug.\n\n- Public endpoint\n- Returns 404 if slug not found",
        "operationId": "get_news_api_v1_news__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g., Validation Error or Duplicate Data)"
          },
          "401": {
            "description": "Unauthorized (Missing or Invalid JWT Token)"
          },
          "403": {
            "description": "Forbidden (Insufficient Permissions or Deactivated Account)"
          },
          "404": {
            "description": "Not Found (The requested resource does not exist)"
          },
          "429": {
            "description": "Too Many Requests (Rate limit exceeded for this endpoint)"
          },
          "500": {
            "description": "Internal Server Error (Unexpected system failure)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/news/{news_id}": {
      "put": {
        "tags": [
          "News"
        ],
        "summary": "Update News",
        "description": "Update an existing news article by ID.\n\n- Admin-only endpoint\n- Partial update (only provided fields)\n- New cover_image replaces the old one",
        "operationId": "update_news_api_v1_news__news_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "news_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "News Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_update_news_api_v1_news__news_id__put"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g., Validation Error or Duplicate Data)"
          },
          "401": {
            "description": "Unauthorized (Missing or Invalid JWT Token)"
          },
          "403": {
            "description": "Forbidden (Insufficient Permissions or Deactivated Account)"
          },
          "404": {
            "description": "Not Found (The requested resource does not exist)"
          },
          "429": {
            "description": "Too Many Requests (Rate limit exceeded for this endpoint)"
          },
          "500": {
            "description": "Internal Server Error (Unexpected system failure)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "News"
        ],
        "summary": "Delete News",
        "description": "Delete a news article by ID.\n\n- Admin-only endpoint\n- Also deletes the cover image from storage",
        "operationId": "delete_news_api_v1_news__news_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "news_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "News Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g., Validation Error or Duplicate Data)"
          },
          "401": {
            "description": "Unauthorized (Missing or Invalid JWT Token)"
          },
          "403": {
            "description": "Forbidden (Insufficient Permissions or Deactivated Account)"
          },
          "404": {
            "description": "Not Found (The requested resource does not exist)"
          },
          "429": {
            "description": "Too Many Requests (Rate limit exceeded for this endpoint)"
          },
          "500": {
            "description": "Internal Server Error (Unexpected system failure)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_news_api_v1_news_post": {
        "properties": {
          "headline": {
            "type": "string",
            "maxLength": 150,
            "title": "Headline"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "source": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "is_featured": {
            "type": "boolean",
            "title": "Is Featured",
            "default": false
          },
          "status": {
            "$ref": "#/components/schemas/ContentStatus",
            "default": "draft"
          },
          "status_reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Reason"
          },
          "cover_image": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Image",
            "description": "Cover image (JPG/PNG/WebP, max 5MB)"
          }
        },
        "type": "object",
        "required": [
          "headline",
          "summary"
        ],
        "title": "Body_create_news_api_v1_news_post"
      },
...
      "Body_update_news_api_v1_news__news_id__put": {
        "properties": {
          "headline": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "source": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "is_featured": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Featured"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContentStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "status_reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Reason"
          },
          "cover_image": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Image",
            "description": "New cover image (replaces existing)"
          },
          "remove_cover_image": {
            "type": "boolean",
            "title": "Remove Cover Image",
            "default": false
          }
        },
        "type": "object",
        "title": "Body_update_news_api_v1_news__news_id__put"
      },
...
      "ContentStatus": {
        "type": "string",
        "enum": [
          "draft",
          "in_review",
          "changes_requested",
          "approved",
          "scheduled",
          "published",
          "unpublished",
          "archived"
        ],
        "title": "ContentStatus",
        "description": "Publication status for all content types."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NewsOut": {
        "properties": {
          "likes_count": {
            "type": "integer",
            "title": "Likes Count",
            "default": 0
          },
          "comments_count": {
            "type": "integer",
            "title": "Comments Count",
            "default": 0
          },
          "is_liked": {
            "type": "boolean",
            "title": "Is Liked",
            "default": false
          },
          "status_changed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Changed At"
          },
          "status_changed_by_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Changed By Id"
          },
          "status_change_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Change Reason"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "headline": {
            "type": "string",
            "title": "Headline"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "cover_image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Image"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "is_featured": {
            "type": "boolean",
            "title": "Is Featured"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "headline",
          "slug",
          "summary",
          "cover_image",
          "source",
          "is_featured",
          "status",
          "published_at",
          "created_at",
          "updated_at"
        ],
        "title": "NewsOut",
        "description": "Response model for a news article."
      },
...
      "PaginatedResponse_NewsOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NewsOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "per_page": {
            "type": "integer",
            "title": "Per Page"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "per_page"
        ],
        "title": "PaginatedResponse[NewsOut]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}